arreglado la seguridad en todo y direcciones bien

+++ app/Http/Controllers/CentroDeportivoController.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace App\Http\Controllers;
+
 use App\Models\centroDeportivo;
 use App\Models\Municipio;
 use Illuminate\Http\Request;
@@ -8,26 +9,31 @@ use Illuminate\Http\Request;
 class CentroDeportivoController extends Controller
 {
 
-     public function listaCentrosDeportivosWelcome()
-     {
+    public function listaCentrosDeportivosWelcome()
+    {
         $centrosDeportivos = CentroDeportivo::all();
-         return view('listaCentrosDeportivosWelcome', compact('centrosDeportivos'));
-     }
+        return view('listaCentrosDeportivosWelcome', compact('centrosDeportivos'));
+    }
 
 
 
     public function listaCentrosDeportivos()
     {
+        if (!auth()->check()) {
+            return redirect()->route('login');
+        }
         $centrosDeportivos = CentroDeportivo::all();
         return view('admin.SuperAdmin.listaCentrosDeportivos', compact('centrosDeportivos'));
-
     }
-    
 
-    public function listaCentrosDeportivosuser(){
+
+    public function listaCentrosDeportivosuser()
+    {
+        if (!auth()->check()) {
+            return redirect()->route('login');
+        }
         $centrosDeportivos = CentroDeportivo::all();
         return view('onclick.nosotros.listaCentrosDeportivosWelcome', compact('centrosDeportivos'));
-
     }
 
     /**
@@ -35,29 +41,35 @@ class CentroDeportivoController extends Controller
      */
     public function create()
     {
+        if (!auth()->check()) {
+            return redirect()->route('login');
+        }
         $municipios = Municipio::all(); // Obtiene todos los municipios
         return view('admin.SuperAdmin.create', compact('municipios')); // Pasa los municipios a la vista
         // return view('admin.SuperAdmin.create');
 
     }
- 
+
     /**
      * Store a newly created resource in storage.
      */
     public function store(Request $request)
     {
+        if (!auth()->check()) {
+            return redirect()->route('login');
+        }
         $request->validate([
-           
+
             'imagen' => 'nullable|image|mimes:jpeg,png,jpg,gif|max:2048',
-           
+
         ]);
-    
+
         $centrosDeportivo = new CentroDeportivo();
-        $centrosDeportivo->nombre = $request->nombre; 
+        $centrosDeportivo->nombre = $request->nombre;
         $centrosDeportivo->direccion = $request->direccion;
         $centrosDeportivo->telefono = $request->telefono;
         $centrosDeportivo->numero_canchas = $request->numero_canchas;
-    
+
         if ($request->hasFile('imagen')) {
             $image = $request->file('imagen');
             $imageName = time() . '.' . $image->getClientOriginalExtension();
@@ -67,20 +79,20 @@ class CentroDeportivoController extends Controller
         $centrosDeportivo->municipio_id = $request->municipio_id;
         $centrosDeportivo->descripcion = $request->descripcion;
         $centrosDeportivo->parqueadero = $request->parqueadero;
-    
+
         // Guardar el objeto en la base de datos
         $centrosDeportivo->save();
-    
+
         // Redireccionar con un mensaje de éxito
         return redirect()->back()->with('success', 'Centro deportivo creado con éxito.');
     }
-    
-    
+
+
     // public function show(string $id)
     // {
     //       // Obtiene el centro deportivo
     // $centroDeportivo = CentroDeportivo::findOrFail($id);
-    
+
     // // Obtiene las canchas asociadas
     // $canchas = $centroDeportivo->canchas;
 
@@ -92,9 +104,13 @@ class CentroDeportivoController extends Controller
      */
     public function edit(string $id)
     {
-    $centroDeportivo = CentroDeportivo::findOrFail($id);
-    $municipios = Municipio::all();
-    return view('admin.SuperAdmin.edit', compact('centroDeportivo', 'municipios'));
+
+        if (!auth()->check()) {
+            return redirect()->route('login');
+        }
+        $centroDeportivo = CentroDeportivo::findOrFail($id);
+        $municipios = Municipio::all();
+        return view('admin.SuperAdmin.edit', compact('centroDeportivo', 'municipios'));
     }
 
     /**
@@ -102,10 +118,13 @@ class CentroDeportivoController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        if (!auth()->check()) {
+            return redirect()->route('login');
+        }
         $request->validate([
             'imagen' => 'nullable|image|mimes:jpeg,png,jpg,gif|max:2048',
         ]);
-    
+
         // Obtener el centro deportivo a actualizar
         $centroDeportivo = CentroDeportivo::findOrFail($id);
         $centroDeportivo->nombre = $request->nombre;
@@ -115,7 +134,7 @@ class CentroDeportivoController extends Controller
         $centroDeportivo->descripcion = $request->descripcion;
         $centroDeportivo->municipio_id = $request->municipio_id;
         $centroDeportivo->parqueadero = $request->parqueadero;
-    
+
         // Actualizar la imagen solo si se subió una nueva
         if ($request->hasFile('imagen')) {
             $image = $request->file('imagen');
@@ -126,7 +145,6 @@ class CentroDeportivoController extends Controller
         $centroDeportivo->save();
 
         return redirect()->route('centroDeportivo.listar')->with('success', 'Centro deportivo actualizado con éxito.');
-    
     }
 
     /**
@@ -134,13 +152,16 @@ class CentroDeportivoController extends Controller
      */
     public function destroy(string $id)
     {
+        if (!auth()->check()) {
+            return redirect()->route('login');
+        }
         $centroDeportivo = CentroDeportivo::findOrFail($id);
-    
+
         // Eliminar la imagen asociada si existe
         if ($centroDeportivo->imagen && file_exists(public_path('img/' . $centroDeportivo->imagen))) {
             unlink(public_path('img/' . $centroDeportivo->imagen));
         }
-        
+
         // Eliminar el centro deportivo
         $centroDeportivo->delete();
